Fix: arreglos testeos para hibernate

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/jpa/UserDaoJpa.java
@@ -175,9 +175,9 @@ public class UserDaoJpa implements UserDao {
 
         final Query query = em.createNativeQuery(sqlQuery);
         @SuppressWarnings("unchecked")
-        BigInteger result = (BigInteger) query.setParameter("jobType", jobType.getValue())
-                .setParameter("id", id).getResultList().stream().findFirst().orElse(BigInteger.valueOf(0));
-        return result.intValue();
+        int result = (int) query.setParameter("jobType", jobType.getValue())
+                .setParameter("id", id).getResultList().stream().findFirst().orElse(0);
+        return result;
     }
 
     private Optional<User> addEncodedImage(Optional<User> maybeUser) {
